judgegrind: valgrind tool for our judging system
Judging based on times() results was quite unsatisfactory because
these times fluctuate wildly especially in the <1s range, giving
random results to the students if they are roughly on the time limit.

Valgrind can help with this since it allows us to precisely track the
instructions executed.

So make a valgrind module with only a few basic capabilities:

1. Count the instructions executed.
2. Guard against any possibility of the student interfering with
   valgrind.
3. Guard against syscall misuse by the students.

Point 2 is very important since it is otherwise quite trivial to
overwrite the internal instruction counter with any number the student
wishes.  We achieve this protection by instrumenting all memory
accesses and checking that the affected memory area is mapped for the
student.

The slight subtlety here is that the student might attempt to trick
valgrind into executing memory accesses whose guards were optimized
out because the access was known to be valid at instrumentation time.
So whenever the student unmaps any memory, and thus loses access
rights, we go the conservative way and tell valgrind to discard all
translations.

Furthermore it is possible to escape valgrind with execve() or via a
client request.  execve() is handled by the syscall filter, and we
just outright disallow client requests when we see them in the IR
stream.

Finally, we are careful to not allow any syscalls that the student
might use to print a fake valgrind-ish result message to valgrind's
output FD, then replace that FD with something else (such as
/dev/null) to redirect the *real* status report.  AFAICS these are
close(), dup2() and dup3().  Fortunately no submission should require
any of those.