judgegrind: award a score
Using the counting facilities in the last commit, I ran some
statistics and determined that the best predictor is the number of
register accesses in the IR (i.e., Put/PutI statements and Get/GetI
expressions).

So use these, slightly simplified, as a score and throw out the rest
of the counting.