Properly guard against the client writing into our log fd
To avoid exploits from the client writing to our log fd, guard all
syscalls operating on fds with a test for the valgrind reserved-fd
boundary.

Unforunately this requires intruding on VG proper: valgrind --log-fd=F
dups F away into the safe range, but then doesn't close it.  To be
sure the user can't write to it, close F if it is outside the std*
range.