judge: remove maybe_instrument_store logic
First, do not leave away instrumentation for non-stack areas, since
those can change and thus require discarding translations in
jg_die_mem, which is very expensive.

This leaves only the stack.  But stack accesses are not constant,
so the maybe_instrument_store logic never worked on them.  Thus we
remove the entire setup surrounding it.

This actually gives a nice speedup.