Replace socket() for java usage
Java insists on using sockets, so killig the program unconditionally on every
socket() call is not an option.

As a band-aid, we still terminate immediately when the program attempts to use
the socket _syscall_, but trap the socket() function within glibc and have it
return EACCESS unconditionally.  This seems to make everyone happy.