judge: count jumps as one reg_instr
We used to not count loops.  This leads to trouble if the optimizer turns the
student's loop into

foo:
  goto foo;

which happened a few times last semester.  Count one reg_instr per jump so that
looping in this way also hits a time limit.

Obviously this will make programs seem slower across the board.  Too bad.