Implement threaded regression tests
We just need to make sure we share the main test queue and all the
output data.

The conditional loading is a bit iffy, but I couldn't find a simpler
solution.  The problem is that 'use threads' acts more magical than
your average module.  So we need to go to some lengths, and stub
share() but not lock() (don't ask!), to avoid warnings both with and
without threads.