regtests: use forks instead of threads
Perl's globbing is currently non-reentrant, so it breaks the
interpreter when running under 'use threads', cf. [1].

The 'forks' module on CPAN works around this issue, since it runs the
perls in separate processes.  This doesn't scale quite so well, but is
a suitable drop-in replacement until such time as 'use threads'
becomes usable for us.

[1] https://rt.perl.org/rt3/Public/Bug/Display.html?id=117823