Use forks instead of threads
The threads support causes errors like these:

  Attempt to free nonexistent shared string '3', Perl interpreter: 0x23cc740 at tests/vg_regtest line 526.

Occasionally followed by an interpreter segfault (!).

I have been advised on #perl that it is "hardly a secret" threads is
broken, and that I should instead 'use forks' as a drop-in
replacement.  So until such time as threads is fixed, let's do that.