{call,cache}grind/tests: use 'rm -f' to clean up
The cleanup is racy: when we get around to running the tests in
parallel, several rm invocations may get the same arguments, causing
one or both to fail (the one that runs unlink() later).

Since no tests currently use the callgrind.out.* files, it is safe to
simply use rm -f to ignore the error.